FE Release 2024-04-16

+++ packages/contracts-communication/contracts/interfaces/IInterchainClientV1.sol
@@ -5,14 +5,14 @@ import {InterchainTxDescriptor} from "../libs/InterchainTransaction.sol";
 
 interface IInterchainClientV1 {
     error InterchainClientV1__FeeAmountTooLow(uint256 actual, uint256 required);
-    error InterchainClientV1__IncorrectDstChainId(uint256 chainId);
+    error InterchainClientV1__IncorrectDstChainId(uint64 chainId);
     error InterchainClientV1__IncorrectMsgValue(uint256 actual, uint256 required);
     error InterchainClientV1__InvalidTransactionVersion(uint16 version);
-    error InterchainClientV1__NoLinkedClient(uint256 chainId);
+    error InterchainClientV1__NoLinkedClient(uint64 chainId);
     error InterchainClientV1__NotEnoughGasSupplied();
     error InterchainClientV1__NotEnoughResponses(uint256 actual, uint256 required);
     error InterchainClientV1__NotEVMClient(bytes32 client);
-    error InterchainClientV1__NotRemoteChainId(uint256 chainId);
+    error InterchainClientV1__NotRemoteChainId(uint64 chainId);
     error InterchainClientV1__TxAlreadyExecuted(bytes32 transactionId);
     error InterchainClientV1__TxNotExecuted(bytes32 transactionId);
     error InterchainClientV1__ZeroReceiver();
@@ -31,7 +31,7 @@ interface IInterchainClientV1 {
      * @param chainId The chain ID for which the client is being set.
      * @param client The address of the client being linked.
      */
-    function setLinkedClient(uint256 chainId, bytes32 client) external;
+    function setLinkedClient(uint64 chainId, bytes32 client) external;
 
     /**
      * @notice Sends a message to another chain via the Interchain Communication Protocol.
@@ -52,7 +52,7 @@ interface IInterchainClientV1 {
      * - entryIndex: the index of the written entry for transaction within the batch.
      */
     function interchainSend(
-        uint256 dstChainId,
+        uint64 dstChainId,
         bytes32 receiver,
         address srcExecutionService,
         address[] calldata srcModules,
@@ -64,7 +64,7 @@ interface IInterchainClientV1 {
         returns (InterchainTxDescriptor memory desc);
 
     function interchainSendEVM(
-        uint256 dstChainId,
+        uint64 dstChainId,
         address receiver,
         address srcExecutionService,
         address[] calldata srcModules,
@@ -99,7 +99,7 @@ interface IInterchainClientV1 {
     /// @param transactionId    The ID of the transaction to write the proof for.
     /// @return dbNonce         The database nonce of the batch containing the written proof for transaction.
     /// @return entryIndex      The index of the written proof for transaction within the batch.
-    function writeExecutionProof(bytes32 transactionId) external returns (uint256 dbNonce, uint64 entryIndex);
+    function writeExecutionProof(bytes32 transactionId) external returns (uint64 dbNonce, uint64 entryIndex);
 
     /**
      * @notice Checks if a transaction is executable.
@@ -120,7 +120,7 @@ interface IInterchainClientV1 {
     /// @param options              Execution options for the message sent, currently gas limit + native gas drop.
     /// @param messageLen           The length of the message being sent.
     function getInterchainFee(
-        uint256 dstChainId,
+        uint64 dstChainId,
         address srcExecutionService,
         address[] calldata srcModules,
         bytes calldata options,
@@ -138,10 +138,10 @@ interface IInterchainClientV1 {
 
     /// @notice Returns the address of the linked client (as bytes32) for a specific chain ID.
     /// @dev Will return 0x0 if no client is linked for the chain ID.
-    function getLinkedClient(uint256 chainId) external view returns (bytes32);
+    function getLinkedClient(uint64 chainId) external view returns (bytes32);
 
     /// @notice Returns the EVM address of the linked client for a specific chain ID.
     /// @dev Will return 0x0 if no client is linked for the chain ID.
     /// Will revert if the client is not an EVM client.
-    function getLinkedClientEVM(uint256 chainId) external view returns (address);
+    function getLinkedClientEVM(uint64 chainId) external view returns (address);
 }
